add dd ocr utils
